requirements.txt
certifi==2019.6.16
DateTime==4.3
dj-database-url==0.5.0
Django==2.2.3
django-crispy-forms==1.7.2
django-filebrowser-no-grappelli==3.7.8
django-heroku==0.3.1
django-sendgrid-v5==0.8.0
django-tinymce4-lite==1.7.5
future==0.17.1
gunicorn==19.9.0
jsmin==2.2.2
Pillow==6.1.0
psycopg2==2.8.3
python-http-client==3.1.0
pytz==2019.1
sendgrid==6.0.5
sqlparse==0.3.0
whitenoise==4.1.3
wincertstore==0.2
zope.interface==4.6.0
